+++ routes/web.php
@@ -1,10 +1,12 @@
 <?php
 
+use App\Http\Controllers\AppealController;
 use App\Http\Controllers\DPAController;
 use App\Http\Controllers\InvestigationController;
 use App\Http\Controllers\LoginController;
 use App\Http\Controllers\ReportController;
 use App\Http\Controllers\UserController;
+use App\Models\Appeal;
 use App\Models\DPA;
 use App\Models\Investigation;
 use App\Models\Report;
@@ -56,3 +58,10 @@ Route::post( '/dpa/new', [ DPAController::class, 'store' ] )->middleware( 'auth'
 Route::get( '/dpa/{dpa}', [ DPAController::class, 'show' ] )->whereAlphaNumeric( 'id' )->middleware( 'auth' )->can( 'view', 'dpa' );
 Route::patch( '/dpa/{dpa}', [ DPAController::class, 'update' ] )->whereAlphaNumeric( 'id' )->middleware( 'auth' )->can( 'update', 'dpa' );
 Route::redirect( '/gdpr', '/dpa' );
+
+/*
+ * Appeal web group
+ */
+Route::get( '/appeals', [ AppealController::class, 'index' ] )->middleware( 'auth' )->can( 'viewAny', Appeal::class );
+Route::get( '/appeal/{appeal}', [ AppealController::class, 'show' ] )->whereNumber( 'id' )->middleware( 'auth' )->can( 'view', 'appeal' );
+Route::patch( '/appeal/{appeal}', [ AppealController::class, 'update' ] )->whereNumber( 'id' )->middleware( 'auth' )->can( 'update', 'appeal' );